Fix tests in launch
