+++ src/test/debugger/extension/configuration/resolvers/base.unit.test.ts
@@ -7,17 +7,15 @@ import { expect } from 'chai';
 import * as path from 'path';
 import * as sinon from 'sinon';
 import { anything, instance, mock, when } from 'ts-mockito';
-import * as typemoq from 'typemoq';
-import { DebugConfiguration, TextDocument, TextEditor, Uri, WorkspaceFolder } from 'vscode';
+import { DebugConfiguration, Uri, WorkspaceFolder } from 'vscode';
 import { CancellationToken } from 'vscode-jsonrpc';
 import { ConfigurationService } from '../../../../../client/common/configuration/service';
-import { PYTHON_LANGUAGE } from '../../../../../client/common/constants';
 import { IConfigurationService } from '../../../../../client/common/types';
 import { BaseConfigurationResolver } from '../../../../../client/debugger/extension/configuration/resolvers/base';
 import { AttachRequestArguments, DebugOptions, LaunchRequestArguments } from '../../../../../client/debugger/types';
 import { IInterpreterService } from '../../../../../client/interpreter/contracts';
 import * as workspaceFolder from '../../../../../client/debugger/extension/configuration/utils/workspaceFolder';
-import * as common from '../../../../../client/debugger/extension/configuration/utils/common';
+import * as helper from '../../../../../client/debugger/extension/configuration/resolvers/helper';
 
 suite('Debugging - Config Resolver', () => {
     class BaseResolver extends BaseConfigurationResolver<AttachRequestArguments | LaunchRequestArguments> {
@@ -39,10 +37,6 @@ suite('Debugging - Config Resolver', () => {
 
         public getWorkspaceFolder(folder: WorkspaceFolder | undefined): Uri | undefined {
             return BaseConfigurationResolver.getWorkspaceFolder(folder);
-        }
-
-        public getProgram(): string | undefined {
-            return BaseConfigurationResolver.getProgram();
         }
 
         public resolveAndUpdatePythonPath(
@@ -73,7 +67,7 @@ suite('Debugging - Config Resolver', () => {
     let interpreterService: IInterpreterService;
     let getWorkspaceFoldersStub: sinon.SinonStub;
     let workspaceStub: sinon.SinonStub;
-    let getActiveTextEditorStub: sinon.SinonStub;
+    let getProgramStub: sinon.SinonStub;
 
     setup(() => {
         configurationService = mock(ConfigurationService);
@@ -81,56 +75,12 @@ suite('Debugging - Config Resolver', () => {
         resolver = new BaseResolver(instance(configurationService), instance(interpreterService));
         getWorkspaceFoldersStub = sinon.stub(workspaceFolder, 'getWorkspaceFolders');
         workspaceStub = sinon.stub(workspaceFolder, 'getWorkspaceFolder');
-        getActiveTextEditorStub = sinon.stub(common, 'getActiveTextEditor');
+        getProgramStub = sinon.stub(helper, 'getProgram');
     });
     teardown(() => {
         sinon.restore();
     });
-    test('Program should return filepath of active editor if file is python', () => {
-        const expectedFileName = 'my.py';
-        const editor = typemoq.Mock.ofType<TextEditor>();
-        const doc = typemoq.Mock.ofType<TextDocument>();
 
-        editor
-            .setup((e) => e.document)
-            .returns(() => doc.object)
-            .verifiable(typemoq.Times.once());
-        doc.setup((d) => d.languageId)
-            .returns(() => PYTHON_LANGUAGE)
-            .verifiable(typemoq.Times.once());
-        doc.setup((d) => d.fileName)
-            .returns(() => expectedFileName)
-            .verifiable(typemoq.Times.once());
-        getActiveTextEditorStub.returns(editor.object);
-
-        const program = resolver.getProgram();
-
-        expect(program).to.be.equal(expectedFileName);
-    });
-    test('Program should return undefined if active file is not python', () => {
-        const editor = typemoq.Mock.ofType<TextEditor>();
-        const doc = typemoq.Mock.ofType<TextDocument>();
-
-        editor
-            .setup((e) => e.document)
-            .returns(() => doc.object)
-            .verifiable(typemoq.Times.once());
-        doc.setup((d) => d.languageId)
-            .returns(() => 'C#')
-            .verifiable(typemoq.Times.once());
-        getActiveTextEditorStub.returns(editor.object);
-
-        const program = resolver.getProgram();
-
-        expect(program).to.be.equal(undefined, 'Not undefined');
-    });
-    test('Program should return undefined if there is no active editor', () => {
-        getActiveTextEditorStub.returns(undefined);
-
-        const program = resolver.getProgram();
-
-        expect(program).to.be.equal(undefined, 'Not undefined');
-    });
     test('Should get workspace folder when workspace folder is provided', () => {
         const expectedUri = Uri.parse('mock');
         const folder: WorkspaceFolder = { index: 0, uri: expectedUri, name: 'mock' };
@@ -149,7 +99,7 @@ suite('Debugging - Config Resolver', () => {
         test(item.title, () => {
             const programPath = path.join('one', 'two', 'three.xyz');
 
-            BaseConfigurationResolver.getProgram = () => programPath;
+            getProgramStub.returns(programPath);
             getWorkspaceFoldersStub.returns(item.workspaceFolders);
 
             const uri = resolver.getWorkspaceFolder(undefined);
@@ -162,7 +112,7 @@ suite('Debugging - Config Resolver', () => {
         const folder: WorkspaceFolder = { index: 0, uri: expectedUri, name: 'mock' };
         const folders: WorkspaceFolder[] = [folder];
 
-        resolver.getProgram = () => undefined;
+        getProgramStub.returns(undefined);
 
         workspaceStub.returns(folder);
 
@@ -178,7 +128,8 @@ suite('Debugging - Config Resolver', () => {
         const folder2: WorkspaceFolder = { index: 1, uri: Uri.parse('134'), name: 'mock2' };
         const folders: WorkspaceFolder[] = [folder1, folder2];
 
-        resolver.getProgram = () => programPath;
+        getProgramStub.returns(programPath);
+
         getWorkspaceFoldersStub.returns(folders);
 
         workspaceStub.returns(folder2);
@@ -193,7 +144,7 @@ suite('Debugging - Config Resolver', () => {
         const folder2: WorkspaceFolder = { index: 1, uri: Uri.parse('134'), name: 'mock2' };
         const folders: WorkspaceFolder[] = [folder1, folder2];
 
-        resolver.getProgram = () => programPath;
+        getProgramStub.returns(programPath);
         getWorkspaceFoldersStub.returns(folders);
 
         workspaceStub.returns(undefined);
